refactor checkout and order request models: change subtotal, total, and shipping cost types from int to double for improved accuracy in calculations

+++ lib/app/data/remote/model/checkout/order_place_request.dart
@@ -8,9 +8,9 @@ String oderPlaceRequestToJson(OderPlaceRequest data) =>
 
 class OderPlaceRequest {
   List<SaleProduct>? saleProducts;
-  int? subTotal;
-  int? total;
-  int? shippingCost;
+  double? subTotal;
+  double? total;
+  double? shippingCost;
   BillingAddress? billingAddress;
   int? paymentMethodId;
   int? customerId;
